refactor: 폴더명 공백문자 제거
